[LCR-142] refactor: DynamoDB 관심자 수 조회 로직 infra 계층으로 이동
